requirements.txt
Django~=4.0.0
djangorestframework~=3.12.4
pydantic~=1.8.2
openpyxl==3.1.1
aiohttp==3.8.4
drf-yasg==1.20.0
